+++ src/main/scala/org/ergoplatform/explorer/protocol/models/ApiDifficulty.scala
@@ -3,7 +3,7 @@ package org.ergoplatform.explorer.protocol.models
 import cats.syntax.either._
 import io.circe.Decoder
 
-final case class ApiDifficulty(value: BigInt)
+final case class ApiDifficulty(value: BigDecimal)
 
 object ApiDifficulty {
 
@@ -11,7 +11,7 @@ object ApiDifficulty {
     Decoder.decodeString.emap { str =>
       Either
         .catchNonFatal {
-          val bInt = BigInt(str)
+          val bInt = BigDecimal(str)
           ApiDifficulty(bInt)
         }
         .leftMap(_ => "Difficulty")
